New Strategies: Nydegger, RevisedDowning, UnnammedStrategy; updated Grofman (from Axelrod first tournament)
